Moved #getAbsoluteFilePath to androidshared to anable logging errors
